+++ src/containers/App.css
@@ -2,7 +2,7 @@
     font-family: 'SEGA LOGO FONT';
     font-style: normal;
     font-weight: normal;
-    src: local('SEGA LOGO FONT'), url('SEGA.woff') format('woff');
+    src: local('SEGA LOGO FONT'), url('..\SEGA.woff') format('woff');
     }
 
 h1{
