+++ src/containers/App.js
@@ -1,8 +1,8 @@
 import React, {Component} from 'react';
-import CardList from './CardList';
-import {robots} from './robots';
-import SearchBox from './SearchBox';
+import CardList from '../components/CardList';
+import SearchBox from '../components/SearchBox';
 import "./App.css";
+import Scroll from '../components/Scroll';
 
 
 
@@ -10,28 +10,39 @@ class App extends Component {
     constructor(){
         super()
         this.state = {
-            robots:robots,
+            robots:[],
             searchfield: ""
         }
+    }
+
+    componentDidMount(){
+        fetch("https://jsonplaceholder.typicode.com/users/")
+            .then(res => { return res.json();})
+            .then(users => {this.setState({robots: users}) })
     }
 
     onSearchChange = (event) => {
         this.setState({
             searchfield:event.target.value
         })
-        
-        
     }
+
+
 
     render(){
         const filtered = this.state.robots.filter(robot => {
             return robot.name.toLowerCase().includes(this.state.searchfield.toLocaleLowerCase())
         })
-        return(
+        if (this.state.robots.length === 0){
+            return <h1>Loading...</h1>
+        }
+        else return(
             <div className="tc">
                 <h1 className="f1">Robofriends</h1>
                 <SearchBox searchChange={this.onSearchChange}/>
+                <Scroll>
                 <CardList robots={filtered}/>
+                </Scroll>
             </div>
         )
     }
